+++ src/server.ts
@@ -2,8 +2,8 @@ import { ReactElement } from 'react';
 import * as ReactDOM from 'react-dom/server';
 
 import { getDataFromTree } from './getDataFromTree';
-
 export * from './getDataFromTree';
+
 export function renderToStringWithData(
   component: ReactElement<any>,
 ): Promise<string> {